+++ modules/editor-service/src/main/java/org/opencastproject/editor/EditorServiceImpl.java
@@ -53,6 +53,7 @@ import org.opencastproject.mediapackage.MediaPackageElementBuilderFactory;
 import org.opencastproject.mediapackage.MediaPackageElementFlavor;
 import org.opencastproject.mediapackage.Publication;
 import org.opencastproject.mediapackage.Track;
+import org.opencastproject.mediapackage.attachment.AttachmentImpl;
 import org.opencastproject.metadata.dublincore.DublinCoreMetadataCollection;
 import org.opencastproject.metadata.dublincore.EventCatalogUIAdapter;
 import org.opencastproject.metadata.dublincore.MetadataJson;
@@ -73,6 +74,7 @@ import org.opencastproject.smil.entity.api.Smil;
 import org.opencastproject.smil.entity.media.api.SmilMediaObject;
 import org.opencastproject.smil.entity.media.container.api.SmilMediaContainer;
 import org.opencastproject.smil.entity.media.element.api.SmilMediaElement;
+import org.opencastproject.util.MimeType;
 import org.opencastproject.util.MimeTypes;
 import org.opencastproject.util.NotFoundException;
 import org.opencastproject.util.data.Tuple;
@@ -88,6 +90,7 @@ import org.opencastproject.workspace.api.Workspace;
 import com.entwinemedia.fn.Fn;
 import com.entwinemedia.fn.data.Opt;
 
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.osgi.service.component.ComponentContext;
@@ -99,13 +102,18 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xml.sax.SAXException;
 
+import java.awt.datatransfer.MimeTypeParseException;
+import java.io.ByteArrayInputStream;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.nio.charset.StandardCharsets;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Base64;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.Dictionary;
@@ -116,6 +124,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 import javax.ws.rs.WebApplicationException;
@@ -158,8 +167,12 @@ public class EditorServiceImpl implements EditorService {
   private String previewTag;
   private String previewSubtype;
   private String waveformSubtype;
+  private String thumbnailSubType;
   private MediaPackageElementFlavor smilSilenceFlavor;
   private ElasticsearchIndex searchIndex;
+  private MediaPackageElementFlavor captionsFlavor;
+  private String thumbnailWfProperty;
+  private List<MediaPackageElementFlavor> thumbnailSourcePrimary;
 
   private static final String DEFAULT_PREVIEW_SUBTYPE = "prepared";
   private static final String DEFAULT_PREVIEW_TAG = "editor";
@@ -168,6 +181,10 @@ public class EditorServiceImpl implements EditorService {
   private static final String DEFAULT_SMIL_CATALOG_TAGS = "archive";
   private static final String DEFAULT_SMIL_SILENCE_FLAVOR = "*/silence";
   private static final String DEFAULT_PREVIEW_VIDEO_SUBTYPE = "video+preview";
+  private static final String DEFAULT_CAPTIONS_FLAVOR = "captions/*";
+  private static final String DEFAULT_THUMBNAIL_SUBTYPE = "player+preview";
+  private static final String DEFAULT_THUMBNAIL_WF_PROPERTY = "thumbnail_edited";
+  private static final List<MediaPackageElementFlavor> DEFAULT_THUMBNAIL_PRIORITY_FLAVOR = new ArrayList<>();
 
   public static final String OPT_PREVIEW_SUBTYPE = "preview.subtype";
   public static final String OPT_PREVIEW_TAG = "preview.tag";
@@ -175,8 +192,11 @@ public class EditorServiceImpl implements EditorService {
   public static final String OPT_SMIL_CATALOG_FLAVOR = "smil.catalog.flavor";
   public static final String OPT_SMIL_CATALOG_TAGS = "smil.catalog.tags";
   public static final String OPT_SMIL_SILENCE_FLAVOR = "smil.silence.flavor";
-
   public static final String OPT_PREVIEW_VIDEO_SUBTYPE = "preview.video.subtype";
+  public static final String OPT_CAPTIONS_FLAVOR = "captions.flavor";
+  public static final String OPT_THUMBNAILSUBTYPE = "thumbnail.subtype";
+  public static final String OPT_THUMBNAIL_WF_PROPERTY = "thumbnail.workflow.property";
+  public static final String OPT_THUMBNAIL_PRIORITY_FLAVOR = "thumbnail.priority.flavor";
 
   private final Set<String> smilCatalogTagSet = new HashSet<>();
 
@@ -253,6 +273,10 @@ public class EditorServiceImpl implements EditorService {
     return waveformSubtype;
   }
 
+  private String getThumbnailSubtype() {
+    return thumbnailSubType;
+  }
+
   @Activate
   @Modified
   public void activate(ComponentContext cc) {
@@ -296,6 +320,27 @@ public class EditorServiceImpl implements EditorService {
     // Preview Video subtype
     previewVideoSubtype =  Objects.toString(properties.get(OPT_PREVIEW_VIDEO_SUBTYPE), DEFAULT_PREVIEW_VIDEO_SUBTYPE);
     logger.debug("Preview video subtype set to '{}'", previewVideoSubtype);
+
+    // Flavor for captions
+    captionsFlavor = MediaPackageElementFlavor.parseFlavor(
+            StringUtils.defaultString((String) properties.get(OPT_CAPTIONS_FLAVOR), DEFAULT_CAPTIONS_FLAVOR));
+    logger.debug("Caption flavor set to '{}'", captionsFlavor);
+
+    thumbnailSubType =  Objects.toString(properties.get(OPT_THUMBNAILSUBTYPE), DEFAULT_THUMBNAIL_SUBTYPE);
+    logger.debug("Thumbnail subtype set to '{}'", thumbnailSubType);
+
+    thumbnailWfProperty = Objects.toString(properties.get(OPT_THUMBNAIL_WF_PROPERTY), DEFAULT_THUMBNAIL_WF_PROPERTY);
+    logger.debug("Thumbnail workflow property set to '{}'", thumbnailWfProperty);
+
+    String thumbnailPriorities = Objects.toString(properties.get(OPT_THUMBNAIL_PRIORITY_FLAVOR));
+    if ("null".equals(thumbnailPriorities)  || thumbnailPriorities.isEmpty()) {
+      thumbnailSourcePrimary = DEFAULT_THUMBNAIL_PRIORITY_FLAVOR;
+    } else {
+      thumbnailSourcePrimary = Arrays.stream(thumbnailPriorities.split(",", -1))
+                                .map(s -> MediaPackageElementFlavor.parseFlavor(s))
+                                .collect(Collectors.toList());
+    }
+    logger.debug("Thumbnail track priority set to '{}'", thumbnailSourcePrimary);
   }
 
   private Boolean elementHasPreviewTag(MediaPackageElement element) {
@@ -435,6 +480,178 @@ public class EditorServiceImpl implements EditorService {
       throw new IOException(e);
     }
 
+  }
+
+  /**
+   * Adds subtitles {@link EditingData.Subtitle} to the media package and sends the updated media package
+   * to the archive. If a subtitle flavor already exists, the subtitle is overwritten
+   *
+   * @param mediaPackage
+   *          the media package to at the SMIL catalog
+   * @param subtitles
+   *          the subtitles to be added
+   * @throws IOException
+   */
+  private void addSubtitleTrack(MediaPackage mediaPackage, List<EditingData.Subtitle> subtitles)
+          throws IOException, IllegalArgumentException {
+    // Check if any of the provided subtitles fail to match the designated flavor
+    for (EditingData.Subtitle subtitle : subtitles) {
+      if (!subtitle.getFlavor().matches(captionsFlavor)) {
+        throw new IllegalArgumentException(
+                "Given subtitle flavor " + subtitle.getFlavor().toString() + " does match caption flavor "
+                        + captionsFlavor);
+      }
+    }
+
+    for (EditingData.Subtitle subtitle : subtitles) {
+      // Generate ID for new tracks
+      String subtitleId = UUID.randomUUID().toString();
+      String trackId = null;
+
+      // Check if subtitle already exists
+      for (Track t : mediaPackage.getTracks()) {
+        if (t.getFlavor().matches(subtitle.getFlavor())) {
+          logger.debug("Set Identifier for Subtitle-Track to: {}", t.getIdentifier());
+          subtitleId = t.getIdentifier();
+          trackId = t.getIdentifier();
+          break;
+        }
+      }
+
+      Track track = mediaPackage.getTrack(trackId);
+
+      // Memorize uri of the previous track file for deletion
+      URI oldTrackURI = null;
+      if (track != null) {
+        oldTrackURI = track.getURI();
+      }
+
+      // Put updated filename in working file repository and update the track.
+      try (InputStream is = IOUtils.toInputStream(subtitle.getSubtitle(), "UTF-8")) {
+        URI subtitleUri = workspace.put(mediaPackage.getIdentifier().toString(), subtitleId, "subtitle.vtt", is);
+
+        // If not exists, create new Track
+        if (track == null) {
+          MediaPackageElementBuilder mpeBuilder = MediaPackageElementBuilderFactory.newInstance().newElementBuilder();
+          track = (Track) mpeBuilder.elementFromURI(subtitleUri, MediaPackageElement.Type.Track, subtitle.getFlavor());
+          mediaPackage.add(track);
+          logger.info("Creating new track for flavor: " + track.getFlavor());
+        }
+
+        track.setURI(subtitleUri);
+        track.setIdentifier(subtitleId);
+        track.setChecksum(null);
+
+        if (oldTrackURI != null && oldTrackURI != subtitleUri) {
+          // Delete the old files from the working file repository and workspace if they were in there
+          logger.info("Removing old track file {}", oldTrackURI);
+          try {
+            workspace.delete(oldTrackURI);
+          } catch (NotFoundException | IOException e) {
+            logger.info("Could not remove track from workspace. Could be it was never there.");
+          }
+        }
+      }
+
+      try {
+        assetManager.takeSnapshot(mediaPackage);
+      } catch (AssetManagerException e) {
+        throw new IOException("Error while adding the updated media package " + mediaPackage + " to the archive", e);
+      }
+    }
+  }
+
+  /**
+   * Adds base64 encoded thumbnail images to the mediapackage and takes a snapshot
+   *
+   * @param editingData
+   *          the editing information
+   * @param mediaPackage
+   *          the media package
+   * @throws MimeTypeParseException
+   * @throws IOException
+   */
+  private void addThumbnailsToArchive(EditingData editingData, MediaPackage mediaPackage)
+          throws MimeTypeParseException, IOException {
+    for (TrackData track : editingData.getTracks()) {
+      String id = track.getId();
+      MediaPackageElementFlavor flavor = new MediaPackageElementFlavor(track.getFlavor().getType(),
+              getThumbnailSubtype());
+      String uri = track.getThumbnailURI();
+
+      // If no uri, what do?
+      if (uri == null || uri.isEmpty()) {
+        continue;
+      }
+      // If uri not base64 encoded, what do?
+      if (!uri.startsWith("data")) {
+        continue;
+      }
+
+      // Decode
+      uri = uri.substring(uri.indexOf(",") + 1);
+      byte[] byteArray;
+      byteArray = Base64.getMimeDecoder().decode(uri);
+      ByteArrayInputStream bais = new ByteArrayInputStream(byteArray);
+
+      // Get MimeType
+      String stringMimeType = detectMimeType(uri);
+      MimeType mimeType = MimeType.mimeType(stringMimeType.split("/")[0], stringMimeType.split("/")[1]);
+
+      // Store image in workspace
+      InputStream inputStream = bais;
+      final String filename = "thumbnail_" + id + "." + mimeType.getSubtype();
+      final String originalThumbnailId = UUID.randomUUID().toString();
+      URI tempThumbnail = null;
+      try {
+        tempThumbnail = workspace
+                .put(mediaPackage.getIdentifier().toString(), originalThumbnailId, filename, inputStream);
+      } catch (IOException e) {
+        throw new IOException("Could not add thumbnail to workspace", e);
+      }
+
+      // Archive
+      final Attachment attachment = AttachmentImpl.fromURI(tempThumbnail);
+      attachment.setFlavor(flavor);
+      attachment.setMimeType(mimeType);
+      Arrays.asList(mediaPackage.getElementsByFlavor(flavor)[0].getTags()).forEach(attachment::addTag);
+      Arrays.stream(mediaPackage.getElementsByFlavor(flavor)).forEach(mediaPackage::remove);
+      mediaPackage.add(attachment);
+
+      // Update publications here in the future?
+
+      // Set workflow property
+      WorkflowPropertiesUtil
+              .storeProperty(assetManager, mediaPackage,
+                      flavor.getType() + "/" + thumbnailWfProperty, "true");
+    }
+
+    // Snapshot
+    // Should we call takeSnapshot just once at the end of setEditData instead?
+    assetManager.takeSnapshot(mediaPackage);
+  }
+
+  /**
+   * Determines if mimetype of a base64 encoded string is one of the listed image mimetypes and returns it.
+   *
+   * @param b64
+   *          the encoded string that is supposed to be an image
+   * @return
+   *          the mimetype
+   * @throws MimeTypeParseException
+   */
+  private String detectMimeType(String b64) throws MimeTypeParseException {
+    var signatures = new HashMap<String, String>();
+    signatures.put("R0lGODdh", "image/gif");
+    signatures.put("iVBORw0KGgo", "image/png");
+    signatures.put("/9j/", "image/jpg");
+
+    for (var s : signatures.entrySet()) {
+      if (b64.indexOf(s.getKey()) == 0) {
+        return s.getValue();
+      }
+    }
+    throw new MimeTypeParseException("No image mimetype found");
   }
 
   private Opt<Publication> getInternalPublication(MediaPackage mp) {
@@ -700,6 +917,21 @@ public class EditorServiceImpl implements EditorService {
       }
     }
 
+    // Get subtitles from the asset manager, so they are guaranteed to be up-to-date after saving
+    Track[] subtitleTracks = mp.getTracks(captionsFlavor);
+    List<EditingData.Subtitle> subtitles = new ArrayList<>();
+    for (Track t: subtitleTracks) {
+      try {
+        File subtitleFile = workspace.get(t.getURI());
+        String subtitleString = FileUtils.readFileToString(subtitleFile, StandardCharsets.UTF_8);
+        subtitles.add(new EditingData.Subtitle(t.getFlavor(), subtitleString));
+      } catch (NotFoundException | IOException e) {
+        errorExit("Could not read subtitle from file", mediaPackageId, ErrorStatus.UNKNOWN);
+      }
+    }
+
+    // Get tracks from the internal publication because it is a lot faster than getting them from the asset manager
+    // for some reason.
     final List<TrackData> tracks = trackList.stream().map(track -> {
       final String uri = signIfNecessary(track.getURI());
       final boolean audioEnabled = !hiddens.contains(tuple(track.getFlavor().getType(), "audio"));
@@ -715,8 +947,17 @@ public class EditorServiceImpl implements EditorService {
       final TrackSubData video = new TrackSubData(track.hasVideo(), videoPreview,
                         videoEnable);
 
+      final String thumbnailURI = Arrays.stream(internalPub.getAttachments())
+              .filter(attachment -> attachment.getFlavor().getType().equals(track.getFlavor().getType()))
+              .filter(attachment -> attachment.getFlavor().getSubtype().equals(getThumbnailSubtype()))
+              .map(MediaPackageElement::getURI).map(this::signIfNecessary)
+              .findAny()
+              .orElse(null);
+
+      final int priority = thumbnailSourcePrimary.indexOf(track.getFlavor());
+
       return new TrackData(track.getFlavor().getType(), track.getFlavor().getSubtype(), audio, video, uri,
-                        track.getIdentifier());
+                        track.getIdentifier(), thumbnailURI, priority);
     }).collect(Collectors.toList());
 
     List<String> waveformList = Arrays.stream(internalPub.getAttachments())
@@ -725,7 +966,7 @@ public class EditorServiceImpl implements EditorService {
             .collect(Collectors.toList());
 
     return new EditingData(segments, tracks, workflows, mp.getDuration(), mp.getTitle(), event.getRecordingStartDate(),
-            event.getSeriesId(), event.getSeriesName(), workflowActive, waveformList);
+            event.getSeriesId(), event.getSeriesName(), workflowActive, waveformList, subtitles);
   }
 
 
@@ -768,7 +1009,8 @@ public class EditorServiceImpl implements EditorService {
   }
 
   @Override
-  public void setEditData(String mediaPackageId, EditingData editingData) throws EditorServiceException {
+  public void setEditData(String mediaPackageId, EditingData editingData) throws EditorServiceException,
+          IOException {
     final Event event = getEvent(mediaPackageId);
 
     if (WorkflowUtil.isActive(event.getWorkflowState())) {
@@ -806,6 +1048,22 @@ public class EditorServiceImpl implements EditorService {
       addSmilToArchive(mediaPackage, smil);
     } catch (IOException e) {
       errorExit("Unable to add SMIL cutting catalog to archive", mediaPackageId, ErrorStatus.UNKNOWN, e);
+    }
+
+    try {
+      addSubtitleTrack(mediaPackage, editingData.getSubtitles());
+    } catch (IOException e) {
+      errorExit("Unable to add subtitle track to archive", mediaPackageId, ErrorStatus.UNKNOWN, e);
+    } catch (IllegalArgumentException e) {
+      errorExit("Illegal subtitle given", mediaPackageId, ErrorStatus.UNKNOWN, e);
+    }
+
+    try {
+      addThumbnailsToArchive(editingData, mediaPackage);
+    } catch (MimeTypeParseException e) {
+      errorExit("Thumbnail had an illegal MimeType", mediaPackageId, ErrorStatus.UNKNOWN, e);
+    } catch (IOException e) {
+      errorExit("Unable to add thumbnail to archive", mediaPackageId, ErrorStatus.UNKNOWN, e);
     }
 
     if (editingData.getPostProcessingWorkflow() != null) {