+++ modules/live-schedule-impl/src/main/java/org/opencastproject/liveschedule/message/SchedulerEventUpdateHandler.java
@@ -20,6 +20,8 @@
  */
 package org.opencastproject.liveschedule.message;
 
+import static org.opencastproject.scheduler.api.SchedulerService.WORKFLOW_CONFIG_PREFIX;
+
 import org.opencastproject.liveschedule.api.LiveScheduleService;
 import org.opencastproject.message.broker.api.scheduler.SchedulerItem;
 import org.opencastproject.message.broker.api.update.SchedulerUpdateHandler;
@@ -38,7 +40,9 @@ import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.Dictionary;
 import java.util.Map;
+import java.util.Objects;
 
 @Component(
     immediate = true,
@@ -51,12 +55,25 @@ public class SchedulerEventUpdateHandler extends UpdateHandler implements Schedu
 
   private static final Logger logger = LoggerFactory.getLogger(SchedulerEventUpdateHandler.class);
 
+  private static final String DESTINATION_SCHEDULER = "SCHEDULER.Liveschedule";
+  private static final String DELETE_ON_CAPTURE_ERROR = "live.deleteOnCapureError";
+
   protected SchedulerService schedulerService;
 
+  private boolean deleteOnCaptureError = true;
+
+  /**
+   * OSGi callback on component activation.
+   *
+   * @param context
+   *          the component context
+   */
   @Activate
   @Override
   public void activate(ComponentContext cc) {
     super.activate(cc);
+    Dictionary properties = cc.getProperties();
+    deleteOnCaptureError = BooleanUtils.toBoolean(Objects.toString(properties.get(DELETE_ON_CAPTURE_ERROR), "true"));
   }
 
   public void execute(final String mpId, final SchedulerItem schedulerItem) {
@@ -77,7 +94,7 @@ public class SchedulerEventUpdateHandler extends UpdateHandler implements Schedu
           break;
         case UpdateProperties:
           // Workflow properties may have been updated (publishLive configuration)
-          String publishLive = schedulerItem.getProperties().get(PUBLISH_LIVE_PROPERTY);
+          String publishLive = schedulerItem.getProperties().get(WORKFLOW_CONFIG_PREFIX.concat(PUBLISH_LIVE_PROPERTY));
           if (publishLive == null) {
             // Not specified so we do nothing. We don't want to delete if we got incomplete props.
             return;
@@ -105,8 +122,8 @@ public class SchedulerEventUpdateHandler extends UpdateHandler implements Schedu
         case UpdateRecordingStatus:
           String state = schedulerItem.getRecordingState();
           if (RecordingState.CAPTURE_FINISHED.equals(state) || RecordingState.UPLOADING.equals(state)
-                  || RecordingState.UPLOADING.equals(state) || RecordingState.CAPTURE_ERROR.equals(state)
-                  || RecordingState.UPLOAD_ERROR.equals(state)) {
+                  || RecordingState.UPLOAD_ERROR.equals(state)
+                  || (RecordingState.CAPTURE_ERROR.equals(state) && deleteOnCaptureError)) {
             if (isLive(mpId)) {
               liveScheduleService.deleteLiveEvent(mpId);
             }